organize requests into namespaces
