+++ src/SpeakEasy.Specifications/Requests/DeleteRequestSpecification.cs
@@ -1,8 +1,9 @@
 using System.Net.Http;
 using Machine.Fakes;
 using Machine.Specifications;
+using SpeakEasy.Requests;
 
-namespace SpeakEasy.Specifications
+namespace SpeakEasy.Specifications.Requests
 {
     [Subject(typeof(DeleteRequest))]
     class DeleteRequestSpecification : WithFakes
