+++ src/SpeakEasy/HttpClient.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using SpeakEasy.Requests;
 
 namespace SpeakEasy
 {
