+++ src/SpeakEasy/Requests/GetLikeRequest.cs
@@ -1,4 +1,4 @@
-namespace SpeakEasy
+namespace SpeakEasy.Requests
 {
     public abstract class GetLikeRequest : HttpRequest
     {
