+++ src/SpeakEasy/Requests/PostRequest.cs
@@ -1,6 +1,6 @@
 using System.Net.Http;
 
-namespace SpeakEasy
+namespace SpeakEasy.Requests
 {
     public sealed class PostRequest : PostLikeRequest
     {
